+++ libft/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: donheo <donheo@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/21 18:14:48 by donheo            #+#    #+#             */
-/*   Updated: 2025/04/17 10:13:59 by donheo           ###   ########.fr       */
+/*   Updated: 2025/04/19 09:16:26 by donheo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		len1;
-	int		len2;
-	int		i;
+	size_t	len1;
+	size_t	len2;
+	size_t	i;
 	char	*new_s;
 
 	len1 = ft_strlen(s1);
